Added QueryBase class to allow querying to be inherited
